requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
black==24.3.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
exceptiongroup==1.2.0
execnet==2.1.0
fasteners==0.19
filelock==3.13.4
h11==0.14.0
idna==3.6
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.0
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
pathspec==0.12.1
pdbp==1.5.0
platformdirs==4.2.0
pluggy==1.4.0
py==1.11.0
Pygments==2.17.2
pynose==1.5.1
pyotp==2.9.0
PySocks==1.7.1
pytest==8.1.1
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.5.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
sbvirtualdisplay==1.3.0
selenium==4.19.0
seleniumbase==4.25.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tabcompleter==1.3.0
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==1.26.18
wsproto==1.2.0
